+++ Tests/Runtime/Tweens/ApplyTests.cs
@@ -5,12 +5,20 @@ using Unity.Entities;
 using Unity.Mathematics;
 using Unity.Transforms;
 
+#if UNITY_TINY_ALL_0_31_0
+using Unity.Tiny;
+#endif
+
+#if UNITY_2D_ENTITIES
+using Unity.U2D.Entities;
+#endif
+
 namespace Timespawn.EntityTween.Tests.Tweens
 {
     public class ApplyTests : TweenTestFixture
     {
         [Test]
-        public void Translation()
+        public void Move()
         {
             Entity entity = EntityManager.CreateEntity();
             Tween.Move(EntityManager, entity, TestStartFloat3, TestEndFloat3, TestDuration);
@@ -27,7 +35,7 @@ namespace Timespawn.EntityTween.Tests.Tweens
         }
 
         [Test]
-        public void Rotation()
+        public void Rotate()
         {
             Entity entity = EntityManager.CreateEntity();
             Tween.Rotate(EntityManager, entity, TestStartQuat, TestEndQuat, TestDuration);
@@ -59,5 +67,25 @@ namespace Timespawn.EntityTween.Tests.Tweens
             NonUniformScale scale = EntityManager.GetComponentData<NonUniformScale>(entity);
             Assert.AreEqual(expected, scale.Value);
         }
+
+#if UNITY_TINY_ALL_0_31_0 || UNITY_2D_ENTITIES
+
+        public void Tint()
+        {
+            Entity entity = EntityManager.CreateEntity();
+            Tween.Tint(EntityManager, entity, TestStartFloat4, TestEndFloat4, TestDuration);
+
+            World.Update();
+            World.Update();
+            OverrideNextDeltaTime(TestDeltaTime);
+            World.Update();
+
+            float percentage = Ease.CalculatePercentage(TestDeltaTime, default, default);
+            float4 expected = math.lerp(TestStartFloat4, TestEndFloat4, percentage);
+            SpriteRenderer renderer = EntityManager.GetComponentData<SpriteRenderer>(entity);
+            Assert.AreEqual(expected, renderer.Color);
+        }
+
+#endif
     }
 }