+++ Tests/Runtime/Tweens/GenerateTests.cs
@@ -4,12 +4,20 @@ using Unity.Entities;
 using Unity.Mathematics;
 using Unity.Transforms;
 
+#if UNITY_TINY_ALL_0_31_0
+using Unity.Tiny;
+#endif
+
+#if UNITY_2D_ENTITIES
+using Unity.U2D.Entities;
+#endif
+
 namespace Timespawn.EntityTween.Tests.Tweens
 {
     public class GenerateTests : TweenTestFixture
     {
         [Test]
-        public void Translation()
+        public void Move()
         {
             Entity entity = EntityManager.CreateEntity();
             Tween.Move(EntityManager, entity, TestStartFloat3, TestEndFloat3, TestDuration);
@@ -17,7 +25,7 @@ namespace Timespawn.EntityTween.Tests.Tweens
         }
 
         [Test]
-        public void Rotation()
+        public void Rotate()
         {
             Entity entity = EntityManager.CreateEntity();
             Tween.Rotate(EntityManager, entity, TestStartQuat, TestEndQuat, TestDuration);
@@ -31,6 +39,18 @@ namespace Timespawn.EntityTween.Tests.Tweens
             Tween.Scale(EntityManager, entity, TestStartFloat3, TestEndFloat3, TestDuration);
             Test<TweenScaleCommand, TweenScale, NonUniformScale, float3>(entity, TestStartFloat3, TestEndFloat3, TestDuration);
         }
+
+#if UNITY_TINY_ALL_0_31_0 || UNITY_2D_ENTITIES
+
+        [Test]
+        public void Tint()
+        {
+            Entity entity = EntityManager.CreateEntity();
+            Tween.Tint(EntityManager, entity, TestStartFloat4, TestEndFloat4, TestDuration);
+            Test<TweenTintCommand, TweenTint, SpriteRenderer, float4>(entity, TestStartFloat4, TestEndFloat4, TestDuration);
+        }
+
+#endif
 
         private void Test<TTweenCommand, TTweenInfo, TTarget, TTweenInfoValue>(Entity entity, TTweenInfoValue start, TTweenInfoValue end, float duration)
             where TTweenCommand : struct, IComponentData, ITweenParams, ITweenInfo<TTweenInfoValue>